refactor: replace field validation functions with Validator and MetadataFieldsHelper class methods
