+++ src/sections/shared/form/DatasetMetadataForm/useSubmitDataset.ts
@@ -6,7 +6,6 @@ import { DatasetRepository } from '../../../../dataset/domain/repositories/Datas
 import { createDataset } from '../../../../dataset/domain/useCases/createDataset'
 import { updateDatasetMetadata } from '../../../../dataset/domain/useCases/updateDatasetMetadata'
 import { MetadataFieldsHelper, type DatasetMetadataFormValues } from './MetadataFieldsHelper'
-import { getValidationFailedFieldError } from '../../../../metadata-block-info/domain/models/fieldValidations'
 import { type DatasetMetadataFormMode } from '.'
 import { QueryParamKey, Route } from '../../../Route.enum'
 import { DatasetNonNumericVersionSearchParam } from '../../../../dataset/domain/models/Dataset'
@@ -74,7 +73,7 @@ export function useSubmitDataset(
         .catch((err) => {
           const errorMessage =
             err instanceof Error && err.message
-              ? getValidationFailedFieldError(err.message) ?? err.message
+              ? MetadataFieldsHelper.getValidationFailedFieldError(err.message) ?? err.message
               : t('validationAlert.content')
 
           setSubmitError(errorMessage)
@@ -104,7 +103,7 @@ export function useSubmitDataset(
         .catch((err) => {
           const errorMessage =
             err instanceof Error && err.message
-              ? getValidationFailedFieldError(err.message) ?? err.message
+              ? MetadataFieldsHelper.getValidationFailedFieldError(err.message) ?? err.message
               : t('validationAlert.content')
 
           setSubmitError(errorMessage)
