+++ src/shared/helpers/Validator.ts
@@ -1,3 +1,5 @@
+type AcceptedProtocols = 'http' | 'https' | 'ftp'
+
 export class Validator {
   static isValidEmail(email: string): boolean {
     const EMAIL_REGEX =
@@ -30,5 +32,42 @@ export class Validator {
   static isValidNumber(input: string): boolean {
     const NUMBER_REGEX = /^\d+$/
     return NUMBER_REGEX.test(input)
+  }
+
+  static isValidURL(url: string, specificProtocols?: AcceptedProtocols[]): boolean {
+    try {
+      const urlObj = new URL(url)
+
+      const acceptedProtocols: AcceptedProtocols[] = ['http', 'https', 'ftp']
+
+      if (!acceptedProtocols.includes(urlObj.protocol.slice(0, -1) as AcceptedProtocols)) {
+        return false
+      }
+
+      if (
+        specificProtocols &&
+        !specificProtocols.includes(urlObj.protocol.slice(0, -1) as AcceptedProtocols)
+      ) {
+        return false
+      }
+
+      if (!this.isValidHostname(urlObj.hostname)) {
+        return false
+      }
+
+      return true
+    } catch (_error) {
+      return false
+    }
+  }
+
+  static isValidHostname(hostname: string): boolean {
+    const hostnameRegex = /^[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/
+    return hostnameRegex.test(hostname)
+  }
+
+  static isValidFloat(value: string): boolean {
+    const FLOAT_REGEX = /^-?\d+(\.\d+)?([eE][+-]?\d+)?$/
+    return FLOAT_REGEX.test(value)
   }
 }
