+++ tests/component/sections/edit-dataset-metadata/EditDatasetMetadata.spec.tsx
@@ -14,6 +14,8 @@ const metadataBlockInfoRepository: MetadataBlockInfoRepository = {} as MetadataB
 const dataset = DatasetMother.createRealistic()
 const metadataBlocksInfoOnEditMode =
   MetadataBlockInfoMother.getByCollectionIdDisplayedOnCreateFalse()
+const metadataBlocksInfoOnCreateMode =
+  MetadataBlockInfoMother.getByCollectionIdDisplayedOnCreateTrue()
 
 describe('EditDatasetMetadata', () => {
   const mountWithDataset = (component: ReactNode, dataset: DatasetModel | undefined) => {
@@ -21,6 +23,9 @@ describe('EditDatasetMetadata', () => {
 
     datasetRepository.getByPersistentId = cy.stub().resolves(dataset)
     metadataBlockInfoRepository.getByCollectionId = cy.stub().resolves(metadataBlocksInfoOnEditMode)
+    metadataBlockInfoRepository.getDisplayedOnCreateByCollectionId = cy
+      .stub()
+      .resolves(metadataBlocksInfoOnCreateMode)
     datasetRepository.updateMetadata = cy.stub().resolves(undefined)
     datasetRepository.getDatasetVersionsSummaries = cy.stub().resolves(undefined)
